feat: load GitHub comment via proxy to hide visitor's IP address (#67)
There is no need to expose the visitors IP address to GitHub. This
happens as soon as the comments are loaded as this is done directly vie
the GitHub API.

This PR introduces a little PHP proxy script to be installed on the
server. Now GitHub sees the server IP only. And we can use a personal
access token to circumvent the rate limiting. The token is available on
the server only.